Game:
changed wordToGuess, wordToDisplay, userInput, isGameOver to public properties with private setter, so they can be used in the tests
Moved usedHelp and guessedLetter to the methods they are used in in order to reduce the scope of these variables
Extracted method HandleLetterGuessCommand from ExecuteCommand to handle the single letters input
Added ToLower in the input in ExecuteCommand, so this will be handled well in the tests

LetterHandler
RevealLetter, HandleLetterGuess
Since wordToDisplay is a property, it cannot be passed by reference. Both are changed to accept string by value and return a string as well.

Player - nothing is done here, but there was a conflict

TestGame - Added tests for some of the methods, still not complete